// src/util/enum.ts
export enum Method {
    POST = "post",
    GET = "get",
    PUT = "put",
    DELETE = "delete",
    PATCH = "patch"
}

export enum UserRole {
    ADMIN = "admin",
    NORMAL = "normal"
}

export enum UserRoleLevel {
    ADMIN = 99999,
    admin = 99999,
    NORMAL = 1,
    normal = 1,
    ANY = 0,
    any = 0
}